refactor(ui): separate power-off flow
